Adjusting namespaces and folder structures
This was done to simplify usages later
