+++ Common/Constants.hpp
@@ -3,4 +3,7 @@
 namespace Argo::Common {
 const int WINDOW_WIDTH = 1024;
 const int WINDOW_HEIGHT = 768;
+const char* const DEFAULT_WINDOW_TITLE = "New Window";
+
+const float TARGET_FPS = 60.0;
 }  // namespace Argo::Common
