+++ Common/FileSystem/CSVHandler.hpp
@@ -1,0 +1,7 @@
+#include "FileHandler.hpp"
+
+namespace Argo::FileSystem {
+
+class CSVHandler : FileHandler {};
+
+}  // namespace Argo::FileSystem
