+++ Common/FileSystem/FileHandler.cpp
@@ -1,0 +1,3 @@
+#include "FileHandler.hpp"
+
+namespace Argo::FileSystem {}
