+++ Common/FileSystem/FileHandler.hpp
@@ -1,0 +1,5 @@
+namespace Argo::FileSystem {
+
+class FileHandler {};
+
+}  // namespace Argo::FileSystem
