+++ Common/FileSystem/IniHandler.cpp
@@ -1,0 +1,3 @@
+#include "IniHandler.hpp"
+
+namespace Argo::FileSystem {}
