+++ Common/FileSystem/IniHandler.hpp
@@ -1,0 +1,7 @@
+#include "FileHandler.hpp"
+
+namespace Argo::FileSystem {
+
+class IniHandler : FileHandler {};
+
+}  // namespace Argo::FileSystem
