+++ Common/FileSystem/JsonHandler.cpp
@@ -1,0 +1,3 @@
+#include "JsonHandler.hpp"
+
+namespace Argo::FileSystem {}
