+++ Common/FileSystem/JsonHandler.hpp
@@ -1,0 +1,7 @@
+#include "FileHandler.hpp"
+
+namespace Argo::FileSystem {
+
+class JsonHandler : FileHandler {};
+
+}  // namespace Argo::FileSystem
