+++ Common/FileSystem/XmlHandler.cpp
@@ -1,0 +1,3 @@
+#include "XmlHandler.hpp"
+
+namespace Argo::FileSystem {}
