+++ Common/FileSystem/XmlHandler.hpp
@@ -1,0 +1,7 @@
+#include "FileHandler.hpp"
+
+namespace Argo::FileSystem {
+
+class XmlHandler : FileHandler {};
+
+}  // namespace Argo::FileSystem
