+++ Common/FileSystem/YamlHandler.cpp
@@ -1,0 +1,3 @@
+#include "YamlHandler.hpp"
+
+namespace Argo::FileSystem {}
