+++ Common/FileSystem/YamlHandler.hpp
@@ -1,0 +1,7 @@
+#include "FileHandler.hpp"
+
+namespace Argo::FileSystem {
+
+class YamlHandler : FileHandler {};
+
+}  // namespace Argo::FileSystem
