--- Common/System/FileSystem/CSVHandler.cpp
@@ -1,3 +1,0 @@
-#include "CSVHandler.hpp"
-
-namespace Argo::Utilities::FileSystem {}
