--- Common/System/FileSystem/CSVHandler.hpp
@@ -1,7 +1,0 @@
-#include "FileHandler.hpp"
-
-namespace Argo::Utilities::FileSystem {
-
-class CSVHandler : FileHandler {};
-
-}  // namespace Argo::Utilities::FileSystem
