--- Common/System/FileSystem/FileHandler.cpp
@@ -1,3 +1,0 @@
-#include "FileHandler.hpp"
-
-namespace Argo::Utilities::FileSystem {}
