--- Common/System/FileSystem/FileHandler.hpp
@@ -1,5 +1,0 @@
-namespace Argo::Utilities::FileSystem {
-
-class FileHandler {};
-
-}  // namespace Argo::Utilities::FileSystem
