--- Common/System/FileSystem/IniHandler.cpp
@@ -1,3 +1,0 @@
-#include "IniHandler.hpp"
-
-namespace Argo::Utilities::FileSystem {}
