--- Common/System/FileSystem/IniHandler.hpp
@@ -1,7 +1,0 @@
-#include "FileHandler.hpp"
-
-namespace Argo::Utilities::FileSystem {
-
-class IniHandler : FileHandler {};
-
-}  // namespace Argo::Utilities::FileSystem
