--- Common/System/FileSystem/JsonHandler.cpp
@@ -1,3 +1,0 @@
-#include "JsonHandler.hpp"
-
-namespace Argo::Utilities::FileSystem {}
