--- Common/System/FileSystem/JsonHandler.hpp
@@ -1,7 +1,0 @@
-#include "FileHandler.hpp"
-
-namespace Argo::Utilities::FileSystem {
-
-class JsonHandler : FileHandler {};
-
-}  // namespace Argo::Utilities::FileSystem
