--- Common/System/FileSystem/XmlHandler.cpp
@@ -1,3 +1,0 @@
-#include "XmlHandler.hpp"
-
-namespace Argo::Utilities::FileSystem {}
