--- Common/System/FileSystem/XmlHandler.hpp
@@ -1,7 +1,0 @@
-#include "FileHandler.hpp"
-
-namespace Argo::Utilities::FileSystem {
-
-class XmlHandler : FileHandler {};
-
-}  // namespace Argo::Utilities::FileSystem
