--- Common/System/FileSystem/YamlHandler.cpp
@@ -1,3 +1,0 @@
-#include "YamlHandler.hpp"
-
-namespace Argo::Utilities::FileSystem {}
