--- Common/System/FileSystem/YamlHandler.hpp
@@ -1,7 +1,0 @@
-#include "FileHandler.hpp"
-
-namespace Argo::Utilities::FileSystem {
-
-class YamlHandler : FileHandler {};
-
-}  // namespace Argo::Utilities::FileSystem
