+++ Common/System/Game.hpp
@@ -13,7 +13,7 @@ class Game : Application {
 
     // Initialize this to 60FPS, so we don't start the first frame with any
     // division by 0 errors or something
-    float deltaTime = 1.0f / 60.0f;
+    float deltaTime = 1.0F / Argo::Common::TARGET_FPS;
 
   private:
     bool isRunning = false;
