--- Networking/Networking.cpp
@@ -1,3 +1,0 @@
-#include "Networking.hpp"
-
-namespace Argo::System {}
