--- Networking/Networking.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-namespace Argo::System {
-class Networking {};
-}  // namespace Argo::System
