+++ Utilities/ColorConverter.cpp
@@ -3,9 +3,9 @@
 namespace Argo::Utilities {
 auto const maxValue = 255.0F;
 
-CommonColor ConvertColor::FromRGBA( Argo::Vec4 rgba ) { return ConvertColor::FromRGBA( &rgba ); }
+CommonColor ConvertColor::FromRGBA( Vec4 rgba ) { return ConvertColor::FromRGBA( &rgba ); }
 
-CommonColor ConvertColor::FromRGBA( Argo::Vec4 *rgba ) {
+CommonColor ConvertColor::FromRGBA( Vec4 *rgba ) {
     auto alpha = rgba->z != NULL ? rgba->z : 1.0F;
     CommonColor color = { rgba->w / maxValue, rgba->x / maxValue, rgba->y / maxValue, alpha };
     return color;
