+++ src/main.cpp
@@ -7,7 +7,7 @@ using namespace Argo::System;
 int main( int /*unused*/, char ** /*unused*/ ) {
     auto game = *new GLGame();
 
-    if ( !game.init( "Game Window", Argo::Common::WINDOW_WIDTH, Argo::Common::WINDOW_HEIGHT ) ) {
+    if ( !game.init( Argo::Common::DEFAULT_WINDOW_TITLE, Argo::Common::WINDOW_WIDTH, Argo::Common::WINDOW_HEIGHT ) ) {
         fprintf( stderr, "Failed to initialize the game. See the logs about for more details." );
         return -1;
     }
